CF1-116 Update entities Check if choice_id is none before making new record.
The prior commits passed the unit tests but caused the program to fail.  This was because `None` was being passed in for choice_id.  Due to the layout of csv files if a faculty member didn't choose a priority for a slot they left that cell empty in the csv.  It made sense when transforming the rows to read that in as `None`.  Since the earlier version of the schema didn't have choice_id as a Primary Key inserting `None` was  acceptable.